Fix Gosplorer allowing to create files on the CD drive

+++ GoOS/GUI/Apps/Gosplorer/MainFrame.cs
@@ -12,7 +12,6 @@ namespace GoOS.GUI.Apps.Gosplorer
     public class MainFrame : Window
     {
         const int IconWidth = 64, IconHeight = 80;
-        
         string Path = @"0:\";
         
         Input AddressBar;
@@ -36,8 +35,7 @@ namespace GoOS.GUI.Apps.Gosplorer
             Closable = true;
             SetDock(WindowDock.Auto);
 
-            AddressBar = new Input(this, 90, 10, (ushort)(Contents.Width - 100 - 20 - 10), 20, "Path") 
-                { Text = Path, Submitted = AddressBar_Submit };
+            AddressBar = new Input(this, 90, 10, (ushort)(Contents.Width - 100 - 20 - 10), 20, "Path") { Text = Path, Submitted = AddressBar_Submit };
             BackButton = new Button(this, 3, 6, 26, 26, string.Empty)
             {
                 Image = arrowleft,
@@ -65,7 +63,7 @@ namespace GoOS.GUI.Apps.Gosplorer
             RefreshButton = new Button(this, 800, 6, 26, 26, string.Empty)
             {
                 Image = refIcon,
-                Clicked = RefreshButton_Click,
+                Clicked = RenderFolderItems,
                 UseSystemStyle = false,
                 RenderWithAlpha = true,
                 BackgroundColour = Color.Transparent
@@ -120,38 +118,39 @@ namespace GoOS.GUI.Apps.Gosplorer
             RefreshButton.Render();
 
             foreach (Button i in Shortcuts) i.Render();
-            foreach (Button i in FolderContents)
-                if (i != null)
-                    i.Render();
+            foreach (Button i in FolderContents) if (i != null) i.Render();
 
             RenderSystemStyleBorder();
         }
 
         private Button GetButtonUnderMouse()
         {
-            foreach (Button i in FolderContents)
-            {
-                if (i.IsMouseOver) { return i; }
-            }
-
+            foreach (Button i in FolderContents) if (i.IsMouseOver) return i;
             return null;
+        }
+
+        private bool IsMouseOverFolderArea
+        {
+            get
+            {
+                return MouseManager.X >= X + 84 &&
+                       MouseManager.X < X + 84 + Contents.Width &&
+                       MouseManager.Y >= Y + 40 &&
+                       MouseManager.Y < Y + 40 + Contents.Height;
+            }
         }
 
         public override void ShowContextMenu()
         {
-            string[] contextMenuEntries = { " New Folder", " New File" };
+            string[] contextMenuEntries = { };
             ContextButton = GetButtonUnderMouse();
 
-            if (ContextButton != null && ContextButton.Image == fileIcon)
-            {
-                contextMenuEntries = new[] { " Open", " Delete" };
-            }
-            else if (ContextButton != null && ContextButton.Image == folderIcon)
-            {
-                contextMenuEntries = new[] { " Open", " Delete" };
-            }
+            if (!IsMouseOverFolderArea) return;
+            else if (ContextButton != null && ContextButton.Image == fileIcon) contextMenuEntries = new[] { " Open", " Delete" };
+            else if (ContextButton != null && ContextButton.Image == folderIcon) contextMenuEntries = new[] { " Open", " Delete" };
+            else if (!Path.StartsWith(@"1:\")) contextMenuEntries = new[] { " New Folder", " New File" };
 
-            ContextMenu.Show(contextMenuEntries, Font_1x.MeasureString(" New Folder "), ContextMenu_Handle);
+            ContextMenu.Show(contextMenuEntries, (ContextButton.Image == fileIcon || ContextButton.Image == folderIcon) ? (ushort)64 : (ushort)96, ContextMenu_Handle);
         }
 
         private void ContextMenu_Handle(string item)
@@ -161,25 +160,22 @@ namespace GoOS.GUI.Apps.Gosplorer
                 case " Open":
                     FolderContents_Clicked(ContextButton.Name);
                     break;
+
                 case " Delete":
-                    if (File.Exists(Path + @"\" + ContextButton.Name))
-                        File.Delete(Path + @"\" + ContextButton.Name);
-                    else if (Directory.Exists(Path + @"\" + ContextButton.Name))
-                        Directory.Delete(Path + @"\" + ContextButton.Name, true);
+                    if (Directory.Exists(Path + @"\" + ContextButton.Name)) Directory.Delete(Path + @"\" + ContextButton.Name, true);
+                    else File.Delete(Path + @"\" + ContextButton.Name);
+
                     RenderFolderItems();
                     break;
+
                 case " New Folder":
                     WindowManager.AddWindow(new NewFolderFrame(Path));
                     break;
+
                 case " New File":
                     WindowManager.AddWindow(new NewFileFrame(Path));
                     break;
             }
-        }
-        
-        private void RefreshButton_Click()
-        {
-            RenderFolderItems();
         }
 
         private void AddressBar_Submit()
